api/pyproject.toml
[project]
name = "api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "agno>=2.3.21",
    "aiosqlite>=0.22.1",
    "fastapi[standard]>=0.128.0",
    "google-genai>=1.56.0",
    "openai>=2.14.0",
    "python-dotenv>=1.2.1",
    "sqlalchemy[asyncio]>=2.0.45",
    "sqlmodel>=0.0.31",
    "mcp>=1.3.0",
    "uvicorn>=0.40.0",
]
